Se movieron todas las clases internas hacia el nivel superior
